separate timeout definition and use it in futex
